Move tests and add end and aligment tests
