+++ test/models/radial_style_gradient_test.dart
@@ -5,7 +5,7 @@ import 'package:flutter_gradient_generator/enums/gradient_style.dart';
 import 'package:flutter_gradient_generator/models/radial_style_gradient.dart';
 import 'package:test/test.dart';
 
-import '../matchers.dart';
+import 'matchers.dart';
 
 void main() {
   group('RadialStyleGradient', () {
@@ -48,6 +48,7 @@ void main() {
         GradientDirection.bottomLeft: Alignment.bottomLeft,
         GradientDirection.bottomCenter: Alignment.bottomCenter,
         GradientDirection.bottomRight: Alignment.bottomRight,
+        GradientDirection.custom(): const Alignment(-0.5, 0),
       };
 
       gradientDirectionToCenterAlignmentMap
